pyproject.toml
[tool.poetry]
name = "elastic-bulk-uploader"
version = "0.1.0"
description = ""
authors = ["Jan Forster <jan.forster@de.ibm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
elasticsearch = "^8.13.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
ipsum = "^0.1.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.9.0"
flake8 = "^7.0.0"
black = "^24.4.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
